requirements.txt
flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
google-auth==2.29.0
google-api-python-client==2.128.0
